agno-pack/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
git+https://github.com/agno-agi/agno.git@v1.4.4#subdirectory=libs/agno
mcp==1.7.1
openai==1.77.0
yfinance==0.2.46
ollama==0.3.3
